#26 - 리팩토링: EOF 추가
깃헙을 위한 리팩토링.
