+++ src/main/java/com/fastcampus/projectboard/controller/MainController.java
@@ -1,0 +1,13 @@
+package com.fastcampus.projectboard.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+    @GetMapping("/")
+    public String root() {
+        return "redirect:/articles";
+    }
+}
